@@ -65,11 +65,12 @@ func GenCrc32(src []byte) uint32 {
 		0x89b8fd09, 0x8d79e0be, 0x803ac667, 0x84fbdbd0,
 		0x9abc8bd5, 0x9e7d9662, 0x933eb0bb, 0x97ffad0c,
 		0xafb010b1, 0xab710d06, 0xa6322bdf, 0xa2f33668,
-		0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4}
+		0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4,
+	}
 
 	j := byte(0)
 	crc32 := uint32(0xFFFFFFFF)
-	for i := 0; i < len(src); i++ {
+	for i := range src {
 		j = (byte(crc32>>24) ^ src[i]) & 0xff
 		crc32 = uint32(uint32(crc32<<8) ^ uint32(crcTable[j]))
 	}
