@@ -2,12 +2,13 @@ package amf
 
 import (
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"io"
 )
 
 // amf0 polymorphic router
-func (d *Decoder) DecodeAmf0(r io.Reader) (interface{}, error) {
+func (d *Decoder) DecodeAmf0(r io.Reader) (any, error) {
 	marker, err := ReadMarker(r)
 	if err != nil {
 		return nil, err
@@ -23,13 +24,13 @@ func (d *Decoder) DecodeAmf0(r io.Reader) (interface{}, error) {
 	case AMF0_OBJECT_MARKER:
 		return d.DecodeAmf0Object(r, false)
 	case AMF0_MOVIECLIP_MARKER:
-		return nil, fmt.Errorf("decode amf0: unsupported type movieclip")
+		return nil, errors.New("decode amf0: unsupported type movieclip")
 	case AMF0_NULL_MARKER:
 		return d.DecodeAmf0Null(r, false)
 	case AMF0_UNDEFINED_MARKER:
 		return d.DecodeAmf0Undefined(r, false)
 	case AMF0_REFERENCE_MARKER:
-		return nil, fmt.Errorf("decode amf0: unsupported type reference")
+		return nil, errors.New("decode amf0: unsupported type reference")
 	case AMF0_ECMA_ARRAY_MARKER:
 		return d.DecodeAmf0EcmaArray(r, false)
 	case AMF0_STRICT_ARRAY_MARKER:
@@ -41,7 +42,7 @@ func (d *Decoder) DecodeAmf0(r io.Reader) (interface{}, error) {
 	case AMF0_UNSUPPORTED_MARKER:
 		return d.DecodeAmf0Unsupported(r, false)
 	case AMF0_RECORDSET_MARKER:
-		return nil, fmt.Errorf("decode amf0: unsupported type recordset")
+		return nil, errors.New("decode amf0: unsupported type recordset")
 	case AMF0_XML_DOCUMENT_MARKER:
 		return d.DecodeAmf0XmlDocument(r, false)
 	case AMF0_TYPED_OBJECT_MARKER:
@@ -62,7 +63,7 @@ func (d *Decoder) DecodeAmf0Number(r io.Reader, decodeMarker bool) (result float
 
 	err = binary.Read(r, binary.BigEndian, &result)
 	if err != nil {
-		return float64(0), fmt.Errorf("amf0 decode: unable to read number: %s", err)
+		return float64(0), fmt.Errorf("amf0 decode: unable to read number: %w", err)
 	}
 
 	return
@@ -80,9 +81,10 @@ func (d *Decoder) DecodeAmf0Boolean(r io.Reader, decodeMarker bool) (result bool
 		return
 	}
 
-	if b == AMF0_BOOLEAN_FALSE {
+	switch b {
+	case AMF0_BOOLEAN_FALSE:
 		return false, nil
-	} else if b == AMF0_BOOLEAN_TRUE {
+	case AMF0_BOOLEAN_TRUE:
 		return true, nil
 	}
 
@@ -101,12 +103,12 @@ func (d *Decoder) DecodeAmf0String(r io.Reader, decodeMarker bool) (result strin
 	var length uint16
 	err = binary.Read(r, binary.BigEndian, &length)
 	if err != nil {
-		return "", fmt.Errorf("decode amf0: unable to decode string length: %s", err)
+		return "", fmt.Errorf("decode amf0: unable to decode string length: %w", err)
 	}
 
-	var bytes = make([]byte, length)
+	bytes := make([]byte, length)
 	if bytes, err = ReadBytes(r, int(length)); err != nil {
-		return "", fmt.Errorf("decode amf0: unable to decode string value: %s", err)
+		return "", fmt.Errorf("decode amf0: unable to decode string value: %w", err)
 	}
 
 	return string(bytes), nil
@@ -132,7 +134,7 @@ func (d *Decoder) DecodeAmf0Object(r io.Reader, decodeMarker bool) (Object, erro
 
 		if key == "" {
 			if err = AssertMarker(r, true, AMF0_OBJECT_END_MARKER); err != nil {
-				return nil, fmt.Errorf("decode amf0: expected object end marker: %s", err)
+				return nil, fmt.Errorf("decode amf0: expected object end marker: %w", err)
 			}
 
 			break
@@ -140,26 +142,25 @@ func (d *Decoder) DecodeAmf0Object(r io.Reader, decodeMarker bool) (Object, erro
 
 		value, err := d.DecodeAmf0(r)
 		if err != nil {
-			return nil, fmt.Errorf("decode amf0: unable to decode object value: %s", err)
+			return nil, fmt.Errorf("decode amf0: unable to decode object value: %w", err)
 		}
 
 		result[key] = value
 	}
 
 	return result, nil
-
 }
 
 // marker: 1 byte 0x05
 // no additional data
-func (d *Decoder) DecodeAmf0Null(r io.Reader, decodeMarker bool) (result interface{}, err error) {
+func (d *Decoder) DecodeAmf0Null(r io.Reader, decodeMarker bool) (result any, err error) {
 	err = AssertMarker(r, decodeMarker, AMF0_NULL_MARKER)
 	return
 }
 
 // marker: 1 byte 0x06
 // no additional data
-func (d *Decoder) DecodeAmf0Undefined(r io.Reader, decodeMarker bool) (result interface{}, err error) {
+func (d *Decoder) DecodeAmf0Undefined(r io.Reader, decodeMarker bool) (result any, err error) {
 	err = AssertMarker(r, decodeMarker, AMF0_UNDEFINED_MARKER)
 	return
 }
@@ -204,12 +205,12 @@ func (d *Decoder) DecodeAmf0EcmaArray(r io.Reader, decodeMarker bool) (Object, e
 	var length uint32
 	err := binary.Read(r, binary.BigEndian, &length)
 	if err != nil {
-		return nil, fmt.Errorf("decode amf0: unable to decode ecma array length: %s", err)
+		return nil, fmt.Errorf("decode amf0: unable to decode ecma array length: %w", err)
 	}
 
 	result, err := d.DecodeAmf0Object(r, false)
 	if err != nil {
-		return nil, fmt.Errorf("decode amf0: unable to decode ecma array object: %s", err)
+		return nil, fmt.Errorf("decode amf0: unable to decode ecma array object: %w", err)
 	}
 
 	return result, nil
@@ -227,15 +228,15 @@ func (d *Decoder) DecodeAmf0StrictArray(r io.Reader, decodeMarker bool) (result 
 	var length uint32
 	err = binary.Read(r, binary.BigEndian, &length)
 	if err != nil {
-		return nil, fmt.Errorf("decode amf0: unable to decode strict array length: %s", err)
+		return nil, fmt.Errorf("decode amf0: unable to decode strict array length: %w", err)
 	}
 
 	d.refCache = append(d.refCache, result)
 
-	for i := uint32(0); i < length; i++ {
+	for range length {
 		tmp, err := d.DecodeAmf0(r)
 		if err != nil {
-			return nil, fmt.Errorf("decode amf0: unable to decode strict array object: %s", err)
+			return nil, fmt.Errorf("decode amf0: unable to decode strict array object: %w", err)
 		}
 		result = append(result, tmp)
 	}
@@ -255,11 +256,11 @@ func (d *Decoder) DecodeAmf0Date(r io.Reader, decodeMarker bool) (result float64
 	}
 
 	if result, err = d.DecodeAmf0Number(r, false); err != nil {
-		return float64(0), fmt.Errorf("decode amf0: unable to decode float in date: %s", err)
+		return float64(0), fmt.Errorf("decode amf0: unable to decode float in date: %w", err)
 	}
 
 	if _, err = ReadBytes(r, 2); err != nil {
-		return float64(0), fmt.Errorf("decode amf0: unable to read 2 trail bytes in date: %s", err)
+		return float64(0), fmt.Errorf("decode amf0: unable to read 2 trail bytes in date: %w", err)
 	}
 
 	return
@@ -277,12 +278,12 @@ func (d *Decoder) DecodeAmf0LongString(r io.Reader, decodeMarker bool) (result s
 	var length uint32
 	err = binary.Read(r, binary.BigEndian, &length)
 	if err != nil {
-		return "", fmt.Errorf("decode amf0: unable to decode long string length: %s", err)
+		return "", fmt.Errorf("decode amf0: unable to decode long string length: %w", err)
 	}
 
-	var bytes = make([]byte, length)
+	bytes := make([]byte, length)
 	if bytes, err = ReadBytes(r, int(length)); err != nil {
-		return "", fmt.Errorf("decode amf0: unable to decode long string value: %s", err)
+		return "", fmt.Errorf("decode amf0: unable to decode long string value: %w", err)
 	}
 
 	return string(bytes), nil
@@ -290,7 +291,7 @@ func (d *Decoder) DecodeAmf0LongString(r io.Reader, decodeMarker bool) (result s
 
 // marker: 1 byte 0x0d
 // no additional data
-func (d *Decoder) DecodeAmf0Unsupported(r io.Reader, decodeMarker bool) (result interface{}, err error) {
+func (d *Decoder) DecodeAmf0Unsupported(r io.Reader, decodeMarker bool) (result any, err error) {
 	err = AssertMarker(r, decodeMarker, AMF0_UNSUPPORTED_MARKER)
 	return
 }
@@ -329,12 +330,12 @@ func (d *Decoder) DecodeAmf0TypedObject(r io.Reader, decodeMarker bool) (TypedOb
 
 	result.Type, err = d.DecodeAmf0String(r, false)
 	if err != nil {
-		return result, fmt.Errorf("decode amf0: typed object unable to determine type: %s", err)
+		return result, fmt.Errorf("decode amf0: typed object unable to determine type: %w", err)
 	}
 
 	result.Object, err = d.DecodeAmf0Object(r, false)
 	if err != nil {
-		return result, fmt.Errorf("decode amf0: typed object unable to determine object: %s", err)
+		return result, fmt.Errorf("decode amf0: typed object unable to determine object: %w", err)
 	}
 
 	return result, nil
