@@ -13,7 +13,7 @@ func (d *Decoder) decodeAbstractMessage(r io.Reader) (result Object, err error) 
 	if err = d.decodeExternal(r, &result,
 		[]string{"body", "clientId", "destination", "headers", "messageId", "timeStamp", "timeToLive"},
 		[]string{"clientIdBytes", "messageIdBytes"}); err != nil {
-		return result, fmt.Errorf("unable to decode abstract external: %s", err)
+		return result, fmt.Errorf("unable to decode abstract external: %w", err)
 	}
 
 	return
@@ -23,14 +23,15 @@ func (d *Decoder) decodeAbstractMessage(r io.Reader) (result Object, err error) 
 func (d *Decoder) decodeAsyncMessageExt(r io.Reader) (result Object, err error) {
 	return d.decodeAsyncMessage(r)
 }
+
 func (d *Decoder) decodeAsyncMessage(r io.Reader) (result Object, err error) {
 	result, err = d.decodeAbstractMessage(r)
 	if err != nil {
-		return result, fmt.Errorf("unable to decode abstract for async: %s", err)
+		return result, fmt.Errorf("unable to decode abstract for async: %w", err)
 	}
 
 	if err = d.decodeExternal(r, &result, []string{"correlationId", "correlationIdBytes"}); err != nil {
-		return result, fmt.Errorf("unable to decode async external: %s", err)
+		return result, fmt.Errorf("unable to decode async external: %w", err)
 	}
 
 	return
@@ -40,24 +41,25 @@ func (d *Decoder) decodeAsyncMessage(r io.Reader) (result Object, err error) {
 func (d *Decoder) decodeAcknowledgeMessageExt(r io.Reader) (result Object, err error) {
 	return d.decodeAcknowledgeMessage(r)
 }
+
 func (d *Decoder) decodeAcknowledgeMessage(r io.Reader) (result Object, err error) {
 	result, err = d.decodeAsyncMessage(r)
 	if err != nil {
-		return result, fmt.Errorf("unable to decode async for ack: %s", err)
+		return result, fmt.Errorf("unable to decode async for ack: %w", err)
 	}
 
 	if err = d.decodeExternal(r, &result); err != nil {
-		return result, fmt.Errorf("unable to decode ack external: %s", err)
+		return result, fmt.Errorf("unable to decode ack external: %w", err)
 	}
 
 	return
 }
 
 // flex.messaging.io.ArrayCollection
-func (d *Decoder) decodeArrayCollection(r io.Reader) (interface{}, error) {
+func (d *Decoder) decodeArrayCollection(r io.Reader) (any, error) {
 	result, err := d.DecodeAmf3(r)
 	if err != nil {
-		return result, fmt.Errorf("cannot decode child of array collection: %s", err)
+		return result, fmt.Errorf("cannot decode child of array collection: %w", err)
 	}
 
 	return result, nil
@@ -70,7 +72,7 @@ func (d *Decoder) decodeExternal(r io.Reader, obj *Object, fieldSets ...[]string
 
 	flagSet, err = readFlags(r)
 	if err != nil {
-		return fmt.Errorf("unable to read flags: %s", err)
+		return fmt.Errorf("unable to read flags: %w", err)
 	}
 
 	for i, flags := range flagSet {
@@ -87,7 +89,14 @@ func (d *Decoder) decodeExternal(r io.Reader, obj *Object, fieldSets ...[]string
 			if (flags & flagBit) != 0 {
 				tmp, err := d.DecodeAmf3(r)
 				if err != nil {
-					return fmt.Errorf("unable to decode external field %s %d %d (%#v): %s", field, i, p, flagSet, err)
+					return fmt.Errorf(
+						"unable to decode external field %s %d %d (%#v): %w",
+						field,
+						i,
+						p,
+						flagSet,
+						err,
+					)
 				}
 				(*obj)[field] = tmp
 			}
@@ -99,7 +108,13 @@ func (d *Decoder) decodeExternal(r io.Reader, obj *Object, fieldSets ...[]string
 					field := fmt.Sprintf("extra_%d_%d", i, j)
 					tmp, err := d.DecodeAmf3(r)
 					if err != nil {
-						return fmt.Errorf("unable to decode post-external field %d %d (%#v): %s", i, j, flagSet, err)
+						return fmt.Errorf(
+							"unable to decode post-external field %d %d (%#v): %w",
+							i,
+							j,
+							flagSet,
+							err,
+						)
 					}
 					(*obj)[field] = tmp
 				}
@@ -107,14 +122,14 @@ func (d *Decoder) decodeExternal(r io.Reader, obj *Object, fieldSets ...[]string
 		}
 	}
 
-	return
+	return err
 }
 
 func readFlags(r io.Reader) (result []uint8, err error) {
 	for {
 		flag, err := ReadByte(r)
 		if err != nil {
-			return result, fmt.Errorf("unable to read flags: %s", err)
+			return result, fmt.Errorf("unable to read flags: %w", err)
 		}
 
 		result = append(result, flag)
