@@ -2,13 +2,14 @@ package amf
 
 import (
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"io"
 	"reflect"
 )
 
 // amf0 polymorphic router
-func (e *Encoder) EncodeAmf0(w io.Writer, val interface{}) (int, error) {
+func (e *Encoder) EncodeAmf0(w io.Writer, val any) (int, error) {
 	if val == nil {
 		return e.EncodeAmf0Null(w, true)
 	}
@@ -37,20 +38,20 @@ func (e *Encoder) EncodeAmf0(w io.Writer, val interface{}) (int, error) {
 	case reflect.Array, reflect.Slice:
 		length := v.Len()
 		arr := make(Array, length)
-		for i := 0; i < length; i++ {
+		for i := range length {
 			arr[i] = v.Index(int(i)).Interface()
 		}
 		return e.EncodeAmf0StrictArray(w, arr, true)
 	case reflect.Map:
 		obj, ok := val.(Object)
 		if !ok {
-			return 0, fmt.Errorf("encode amf0: unable to create object from map")
+			return 0, errors.New("encode amf0: unable to create object from map")
 		}
 		return e.EncodeAmf0Object(w, obj, true)
 	}
 
 	if _, ok := val.(TypedObject); ok {
-		return 0, fmt.Errorf("encode amf0: unsupported type typed object")
+		return 0, errors.New("encode amf0: unsupported type typed object")
 	}
 
 	return 0, fmt.Errorf("encode amf0: unsupported type %s", v.Type())
@@ -77,7 +78,7 @@ func (e *Encoder) EncodeAmf0Number(w io.Writer, val float64, encodeMarker bool) 
 
 // marker: 1 byte 0x01
 // format: 1 byte, 0x00 = false, 0x01 = true
-func (e *Encoder) EncodeAmf0Boolean(w io.Writer, val bool, encodeMarker bool) (n int, err error) {
+func (e *Encoder) EncodeAmf0Boolean(w io.Writer, val, encodeMarker bool) (n int, err error) {
 	if encodeMarker {
 		if err = WriteMarker(w, AMF0_BOOLEAN_MARKER); err != nil {
 			return
@@ -118,13 +119,13 @@ func (e *Encoder) EncodeAmf0String(w io.Writer, val string, encodeMarker bool) (
 	length := uint16(len(val))
 	err = binary.Write(w, binary.BigEndian, length)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode string length: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode string length: %w", err)
 	}
 	n += 2
 
 	m, err = w.Write([]byte(val))
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode string value: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode string value: %w", err)
 	}
 	n += m
 
@@ -138,7 +139,7 @@ func (e *Encoder) EncodeAmf0String(w io.Writer, val string, encodeMarker bool) (
 func (e *Encoder) EncodeAmf0Object(w io.Writer, val Object, encodeMarker bool) (n int, err error) {
 	if encodeMarker {
 		if err = WriteMarker(w, AMF0_OBJECT_MARKER); err != nil {
-			return
+			return n, err
 		}
 		n += 1
 	}
@@ -147,30 +148,30 @@ func (e *Encoder) EncodeAmf0Object(w io.Writer, val Object, encodeMarker bool) (
 	for k, v := range val {
 		m, err = e.EncodeAmf0String(w, k, false)
 		if err != nil {
-			return n, fmt.Errorf("encode amf0: unable to encode object key: %s", err)
+			return n, fmt.Errorf("encode amf0: unable to encode object key: %w", err)
 		}
 		n += m
 
 		m, err = e.EncodeAmf0(w, v)
 		if err != nil {
-			return n, fmt.Errorf("encode amf0: unable to encode object value: %s", err)
+			return n, fmt.Errorf("encode amf0: unable to encode object value: %w", err)
 		}
 		n += m
 	}
 
 	m, err = e.EncodeAmf0String(w, "", false)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode object empty string: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode object empty string: %w", err)
 	}
 	n += m
 
 	err = WriteMarker(w, AMF0_OBJECT_END_MARKER)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to object end marker: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to object end marker: %w", err)
 	}
 	n += 1
 
-	return
+	return n, err
 }
 
 // marker: 1 byte 0x05
@@ -205,7 +206,11 @@ func (e *Encoder) EncodeAmf0Undefined(w io.Writer, encodeMarker bool) (n int, er
 // - normal object format:
 //   - loop encoded string followed by encoded value
 //   - terminated with empty string followed by 1 byte 0x09
-func (e *Encoder) EncodeAmf0EcmaArray(w io.Writer, val Object, encodeMarker bool) (n int, err error) {
+func (e *Encoder) EncodeAmf0EcmaArray(
+	w io.Writer,
+	val Object,
+	encodeMarker bool,
+) (n int, err error) {
 	if encodeMarker {
 		if err = WriteMarker(w, AMF0_ECMA_ARRAY_MARKER); err != nil {
 			return
@@ -217,13 +222,13 @@ func (e *Encoder) EncodeAmf0EcmaArray(w io.Writer, val Object, encodeMarker bool
 	length := uint32(len(val))
 	err = binary.Write(w, binary.BigEndian, length)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode ecma array length: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode ecma array length: %w", err)
 	}
 	n += 4
 
 	m, err = e.EncodeAmf0Object(w, val, false)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode ecma array object: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode ecma array object: %w", err)
 	}
 	n += m
 
@@ -234,7 +239,11 @@ func (e *Encoder) EncodeAmf0EcmaArray(w io.Writer, val Object, encodeMarker bool
 // format:
 // - 4 byte big endian uint32 to determine length of associative array
 // - n (length) encoded values
-func (e *Encoder) EncodeAmf0StrictArray(w io.Writer, val Array, encodeMarker bool) (n int, err error) {
+func (e *Encoder) EncodeAmf0StrictArray(
+	w io.Writer,
+	val Array,
+	encodeMarker bool,
+) (n int, err error) {
 	if encodeMarker {
 		if err = WriteMarker(w, AMF0_STRICT_ARRAY_MARKER); err != nil {
 			return
@@ -246,14 +255,14 @@ func (e *Encoder) EncodeAmf0StrictArray(w io.Writer, val Array, encodeMarker boo
 	length := uint32(len(val))
 	err = binary.Write(w, binary.BigEndian, length)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode strict array length: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode strict array length: %w", err)
 	}
 	n += 4
 
 	for _, v := range val {
 		m, err = e.EncodeAmf0(w, v)
 		if err != nil {
-			return n, fmt.Errorf("encode amf0: unable to encode strict array element: %s", err)
+			return n, fmt.Errorf("encode amf0: unable to encode strict array element: %w", err)
 		}
 		n += m
 	}
@@ -265,7 +274,11 @@ func (e *Encoder) EncodeAmf0StrictArray(w io.Writer, val Array, encodeMarker boo
 // format:
 // - 4 byte big endian uint32 header to determine size
 // - n (size) byte utf8 string
-func (e *Encoder) EncodeAmf0LongString(w io.Writer, val string, encodeMarker bool) (n int, err error) {
+func (e *Encoder) EncodeAmf0LongString(
+	w io.Writer,
+	val string,
+	encodeMarker bool,
+) (n int, err error) {
 	if encodeMarker {
 		if err = WriteMarker(w, AMF0_LONG_STRING_MARKER); err != nil {
 			return
@@ -277,13 +290,13 @@ func (e *Encoder) EncodeAmf0LongString(w io.Writer, val string, encodeMarker boo
 	length := uint32(len(val))
 	err = binary.Write(w, binary.BigEndian, length)
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode long string length: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode long string length: %w", err)
 	}
 	n += 4
 
 	m, err = w.Write([]byte(val))
 	if err != nil {
-		return n, fmt.Errorf("encode amf0: unable to encode long string value: %s", err)
+		return n, fmt.Errorf("encode amf0: unable to encode long string value: %w", err)
 	}
 	n += m
 
