@@ -80,7 +80,23 @@ func TestEncodeAmf0String(t *testing.T) {
 
 func TestEncodeAmf0Object(t *testing.T) {
 	buf := new(bytes.Buffer)
-	expect := []byte{0x03, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x02, 0x00, 0x03, 0x62, 0x61, 0x72, 0x00, 0x00, 0x09}
+	expect := []byte{
+		0x03,
+		0x00,
+		0x03,
+		0x66,
+		0x6f,
+		0x6f,
+		0x02,
+		0x00,
+		0x03,
+		0x62,
+		0x61,
+		0x72,
+		0x00,
+		0x00,
+		0x09,
+	}
 
 	enc := new(Encoder)
 
@@ -101,7 +117,27 @@ func TestEncodeAmf0Object(t *testing.T) {
 
 func TestEncodeAmf0EcmaArray(t *testing.T) {
 	buf := new(bytes.Buffer)
-	expect := []byte{0x08, 0x00, 0x00, 0x00, 0x01, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x02, 0x00, 0x03, 0x62, 0x61, 0x72, 0x00, 0x00, 0x09}
+	expect := []byte{
+		0x08,
+		0x00,
+		0x00,
+		0x00,
+		0x01,
+		0x00,
+		0x03,
+		0x66,
+		0x6f,
+		0x6f,
+		0x02,
+		0x00,
+		0x03,
+		0x62,
+		0x61,
+		0x72,
+		0x00,
+		0x00,
+		0x09,
+	}
 
 	enc := new(Encoder)
 
@@ -120,7 +156,29 @@ func TestEncodeAmf0EcmaArray(t *testing.T) {
 
 func TestEncodeAmf0StrictArray(t *testing.T) {
 	buf := new(bytes.Buffer)
-	expect := []byte{0x0a, 0x00, 0x00, 0x00, 0x03, 0x00, 0x40, 0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x05}
+	expect := []byte{
+		0x0a,
+		0x00,
+		0x00,
+		0x00,
+		0x03,
+		0x00,
+		0x40,
+		0x14,
+		0x00,
+		0x00,
+		0x00,
+		0x00,
+		0x00,
+		0x00,
+		0x02,
+		0x00,
+		0x03,
+		0x66,
+		0x6f,
+		0x6f,
+		0x05,
+	}
 
 	enc := new(Encoder)
 
@@ -163,7 +221,7 @@ func TestEncodeAmf0LongString(t *testing.T) {
 	testBytes := []byte("12345678")
 
 	tbuf := new(bytes.Buffer)
-	for i := 0; i < 65536; i++ {
+	for range 65536 {
 		tbuf.Write(testBytes)
 	}
 
@@ -198,7 +256,12 @@ func TestEncodeAmf0LongString(t *testing.T) {
 	for buf.Len() > 0 {
 		n, err := buf.Read(tmpBuf)
 		if err != nil {
-			t.Fatalf("test long string result check, read data(%d) error: %s, n: %d", counter, err, n)
+			t.Fatalf(
+				"test long string result check, read data(%d) error: %s, n: %d",
+				counter,
+				err,
+				n,
+			)
 		}
 		if n != 8 {
 			t.Fatalf("test long string result check, read data(%d) n: %d", counter, n)
