@@ -150,7 +150,25 @@ func TestDecodeAmf0String(t *testing.T) {
 }
 
 func TestDecodeAmf0Object(t *testing.T) {
-	buf := bytes.NewReader([]byte{0x03, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x02, 0x00, 0x03, 0x62, 0x61, 0x72, 0x00, 0x00, 0x09})
+	buf := bytes.NewReader(
+		[]byte{
+			0x03,
+			0x00,
+			0x03,
+			0x66,
+			0x6f,
+			0x6f,
+			0x02,
+			0x00,
+			0x03,
+			0x62,
+			0x61,
+			0x72,
+			0x00,
+			0x00,
+			0x09,
+		},
+	)
 
 	dec := &Decoder{}
 
@@ -269,7 +287,29 @@ func TestDecodeReference(t *testing.T) {
 */
 
 func TestDecodeAmf0EcmaArray(t *testing.T) {
-	buf := bytes.NewReader([]byte{0x08, 0x00, 0x00, 0x00, 0x01, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x02, 0x00, 0x03, 0x62, 0x61, 0x72, 0x00, 0x00, 0x09})
+	buf := bytes.NewReader(
+		[]byte{
+			0x08,
+			0x00,
+			0x00,
+			0x00,
+			0x01,
+			0x00,
+			0x03,
+			0x66,
+			0x6f,
+			0x6f,
+			0x02,
+			0x00,
+			0x03,
+			0x62,
+			0x61,
+			0x72,
+			0x00,
+			0x00,
+			0x09,
+		},
+	)
 
 	dec := &Decoder{}
 
@@ -316,7 +356,31 @@ func TestDecodeAmf0EcmaArray(t *testing.T) {
 }
 
 func TestDecodeAmf0StrictArray(t *testing.T) {
-	buf := bytes.NewReader([]byte{0x0a, 0x00, 0x00, 0x00, 0x03, 0x00, 0x40, 0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x03, 0x66, 0x6f, 0x6f, 0x05})
+	buf := bytes.NewReader(
+		[]byte{
+			0x0a,
+			0x00,
+			0x00,
+			0x00,
+			0x03,
+			0x00,
+			0x40,
+			0x14,
+			0x00,
+			0x00,
+			0x00,
+			0x00,
+			0x00,
+			0x00,
+			0x02,
+			0x00,
+			0x03,
+			0x66,
+			0x6f,
+			0x6f,
+			0x05,
+		},
+	)
 
 	dec := &Decoder{}
 
@@ -514,7 +578,6 @@ func TestDecodeAmf0XmlDocument(t *testing.T) {
 }
 
 func TestDecodeAmf0TypedObject(t *testing.T) {
-
 	buf := bytes.NewReader([]byte{
 		0x10, 0x00, 0x0F, 'o', 'r', 'g',
 		'.', 'a', 'm', 'f', '.', 'A',
