@@ -165,7 +165,8 @@ func TestDecodeAmf3String(t *testing.T) {
 }
 
 func TestDecodeAmf3Array(t *testing.T) {
-	buf := bytes.NewReader([]byte{0x09, 0x13, 0x01,
+	buf := bytes.NewReader([]byte{
+		0x09, 0x13, 0x01,
 		0x06, 0x03, '1',
 		0x06, 0x03, '2',
 		0x06, 0x03, '3',
