@@ -2,6 +2,7 @@ package amf
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 )
 
@@ -45,7 +46,7 @@ func MetaDataReform(p []byte, flag uint8) ([]byte, error) {
 				p = b
 			}
 		default:
-			return nil, fmt.Errorf("setFrameFrame error")
+			return nil, errors.New("setFrameFrame error")
 		}
 	case DEL:
 		v, err := decoder.Decode(r, AMF0)
@@ -58,7 +59,7 @@ func MetaDataReform(p []byte, flag uint8) ([]byte, error) {
 				p = p[len(setFrameFrame):]
 			}
 		default:
-			return nil, fmt.Errorf("metadata error")
+			return nil, errors.New("metadata error")
 		}
 	default:
 		return nil, fmt.Errorf("invalid flag:%d", flag)
