@@ -1,15 +1,14 @@
 package httpflv
 
 import (
+	"context"
 	"errors"
 	"io"
 	"sync"
-	"sync/atomic"
 
 	"github.com/zijiren233/livelib/av"
 	"github.com/zijiren233/livelib/container/flv"
 	"github.com/zijiren233/livelib/protocol/amf"
-	"github.com/zijiren233/livelib/utils"
 	"github.com/zijiren233/stream"
 )
 
@@ -19,7 +18,6 @@ const (
 )
 
 type HttpFlvWriter struct {
-	t         utils.Timestamp
 	headerBuf []byte
 	w         *stream.Writer
 	inited    bool
@@ -27,8 +25,8 @@ type HttpFlvWriter struct {
 
 	packetQueue chan *av.Packet
 
-	closed uint32
-	wg     sync.WaitGroup
+	closed bool
+	mu     sync.RWMutex
 }
 
 type HttpFlvWriterConf func(*HttpFlvWriter)
@@ -56,74 +54,79 @@ func NewHttpFLVWriter(w io.Writer, conf ...HttpFlvWriterConf) *HttpFlvWriter {
 }
 
 func (w *HttpFlvWriter) Write(p *av.Packet) (err error) {
-	w.wg.Add(1)
-	defer w.wg.Done()
-
-	if w.Closed() {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+	if w.closed {
 		return av.ErrClosed
 	}
-	p = p.Clone()
-	p.TimeStamp = w.t.RecTimeStamp(p.TimeStamp, p.First)
-	select {
-	case w.packetQueue <- p:
-	default:
-		av.DropPacket(w.packetQueue)
+
+	for {
+		select {
+		case w.packetQueue <- p:
+			return
+		default:
+			av.DropPacket(w.packetQueue)
+		}
 	}
-	return
 }
 
-func (w *HttpFlvWriter) SendPacket() error {
+func (w *HttpFlvWriter) SendPacket(ctx context.Context) error {
 	var typeID uint8
-	for p := range w.packetQueue {
-		if !w.inited {
-			if err := w.w.Bytes(flv.FlvFirstHeader).Error(); err != nil {
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case p, ok := <-w.packetQueue:
+			if !ok {
+				return nil
+			}
+			if !w.inited {
+				if err := w.w.Bytes(flv.FlvFirstHeader).Error(); err != nil {
+					return err
+				}
+				w.inited = true
+			}
+
+			if p.IsVideo {
+				typeID = av.TAG_VIDEO
+			} else if p.IsMetadata {
+				var err error
+				typeID = av.TAG_SCRIPTDATAAMF0
+				p = p.DeepClone()
+				p.Data, err = amf.MetaDataReform(p.Data, amf.DEL)
+				if err != nil {
+					return err
+				}
+			} else if p.IsAudio {
+				typeID = av.TAG_AUDIO
+			} else {
+				return errors.New("not allowed packet type")
+			}
+			dataLen := len(p.Data)
+			preDataLen := dataLen + headerLen
+			timestampExt := p.TimeStamp >> 24
+
+			if err := w.w.
+				U8(typeID).
+				U24(uint32(dataLen)).
+				U24(p.TimeStamp).
+				U8(uint8(timestampExt)).
+				U24(0).
+				Bytes(p.Data).
+				U32(uint32(preDataLen)).Error(); err != nil {
 				return err
 			}
-			w.inited = true
-		}
-
-		if p.IsVideo {
-			typeID = av.TAG_VIDEO
-		} else if p.IsMetadata {
-			var err error
-			typeID = av.TAG_SCRIPTDATAAMF0
-			p = p.DeepClone()
-			p.Data, err = amf.MetaDataReform(p.Data, amf.DEL)
-			if err != nil {
-				return err
-			}
-		} else if p.IsAudio {
-			typeID = av.TAG_AUDIO
-		} else {
-			return errors.New("not allowed packet type")
-		}
-		dataLen := len(p.Data)
-		preDataLen := dataLen + headerLen
-		timestampExt := p.TimeStamp >> 24
-
-		if err := w.w.
-			U8(typeID).
-			U24(uint32(dataLen)).
-			U24(p.TimeStamp).
-			U8(uint8(timestampExt)).
-			U24(0).
-			Bytes(p.Data).
-			U32(uint32(preDataLen)).Error(); err != nil {
-			return err
 		}
 	}
-	return nil
 }
 
 func (w *HttpFlvWriter) Close() error {
-	if !atomic.CompareAndSwapUint32(&w.closed, 0, 1) {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+	if w.closed {
 		return av.ErrClosed
 	}
-	w.wg.Wait()
+	w.closed = true
 	close(w.packetQueue)
 	return nil
-}
-
-func (w *HttpFlvWriter) Closed() bool {
-	return atomic.LoadUint32(&w.closed) == 1
 }
